% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_file_name.R
\name{assemble_file_name}
\alias{assemble_file_name}
\title{Construct a time-stamped file-name}
\usage{
assemble_file_name(prefix, suffix = "xlsx", dt_chr = "", hr_chr = "")
}
\arguments{
\item{prefix}{\if{html}{\out{<chr>}} initial part of file-name}

\item{suffix}{\if{html}{\out{<chr>}} file extension}

\item{dt_chr}{\if{html}{\out{<chr>}} date as "yyyy-mm-dd" or other format}

\item{hr_chr}{\if{html}{\out{<chr>}} time as "hhmm" or other format}
}
\value{
\if{html}{\out{<chr>}} file-name
}
\description{
Construct a file-name following a time-stamp naming convention.
}
\examples{
assemble_file_name(prefix = "my_xl_data")
assemble_file_name(prefix = "my-tsv-data", suffix = "tsv")
assemble_file_name(prefix = "initial-data", dt_chr = "2019-01-31")
assemble_file_name(prefix = "current-xl-data", dt_chr = "2024-01-31", hr_chr = "1330")

}
