% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xl_data_hdr.R
\name{make_hdr_range}
\alias{make_hdr_range}
\title{Excel range of columns in a single row}
\usage{
make_hdr_range(col_start = "B", col_stop = "B", row_hdr = 1L)
}
\arguments{
\item{col_start}{\if{html}{\out{<chr>}} initial column}

\item{col_stop}{\if{html}{\out{<chr>}} final column}

\item{row_hdr}{\if{html}{\out{<int>}} desired row, e.g., containing column names}
}
\value{
\if{html}{\out{<chr>}} single string of the form "A1:B2"
}
\description{
Given a pair of column end-points in a single row
return the corresponding Excel range.  Can be used
to extract column names.
}
\examples{
make_hdr_range("A", "Z")
make_hdr_range("A", "Z", 2)

}
