% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xl_data_hdr.R
\name{make_xl_range_vec}
\alias{make_xl_range_vec}
\title{Construct Excel ranges from end-points}
\usage{
make_xl_range_vec(
  col_start = "B",
  col_stop = "B",
  row_start = 6L,
  row_stop = 6L
)
}
\arguments{
\item{col_start}{\if{html}{\out{<chr>}} initial column}

\item{col_stop}{\if{html}{\out{<chr>}} final column}

\item{row_start}{\if{html}{\out{<int>}} initial row}

\item{row_stop}{\if{html}{\out{<int>}} final row}
}
\value{
\if{html}{\out{<chr>}} vector of elements of the form "A1:B2"
}
\description{
Given a set of Excel end-points, retrun the corresponding
Excel ranges
}
\examples{
make_xl_range_vec("A", "B", 1, 2)
make_xl_range_vec(c("A", "B"), c("B", "C"), 1, 2)

}
