% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_columns.R
\name{rename_columns}
\alias{rename_columns}
\title{Rename the columns of a data frame}
\usage{
rename_columns(df, lookup_tbl, old_names, new_names)
}
\arguments{
\item{df}{\if{html}{\out{<tbl>}} data frame whose columns are to be renamed}

\item{lookup_tbl}{\if{html}{\out{<tbl>}} tibble containing (old, new) pairs of names}

\item{old_names}{\if{html}{\out{<idx>}} column name or index of old names in lookup_tbl}

\item{new_names}{\if{html}{\out{<idx>}} column name or index of new names in lookup_tbl}
}
\value{
\if{html}{\out{<df>}} data frame with new column names
}
\description{
The solution presented here was given in
"https://stackoverflow.com/a/43742442/3927208"
}
\examples{
\dontrun{
  df_tst <- tibble(
    A = 1:3,
    B = 2L * A,
    C = 3L * A)
  lu_tst <- tibble(
    old = LETTERS[1:3],
    new = letters[1:3])
  new_df <- rename_columns(df_tst, lu_tst, "old", "new")
}

}
